Cone Radius Calculation (Vectorized using map)

This function calculates the radius of a cone for each height value in a numeric vector `x_vals`. It uses the `map_dbl` function from the `purrr` package.

Usage

cone_radius_map(x_vals)

Arguments

x_vals

A numeric vector of input height values for which the cone radius will be calculated.

Value

A numeric vector containing the calculated cone radius corresponding to each input height value in `x_vals`.

Author

Lodrik Adam

Examples

cone_radius_map(c(-1, 4, 8, 8 + pi / 4, 10))
#> [1] 0.00000 0.50000 1.00000 2.06066 0.00000
cone_radius_map(c(-5, 15))
#> [1] 0 0
cone_radius_map(seq(0, 10, by = 2))
#> [1] 0.00 0.25 0.50 0.75 1.00 0.00